Black magic in data controller to handle annotation items
In the XMP, items are stored in a dict, because it makes sense to arrange
them in category. On the data, their type has no immediate impact on the
controller or the view, so a list of items is more suitable.
Here, we used "model" as the original dict, and "local_model" the list
containing all pointers from the relevant categories of the model.
As we store pointers in local_model, some changes can be only performed on
it to be immediately transmitted to the original model. For bigger
transformation (addition, deletion, type change), some manipulations are
inevitable