Move all annotation_controller and annotation_interface logic in qidatawidget
This reduces the number of classes and gives more genericity to QiDataWidget
Also was moved the annotator's selection tool into QiDataWidget. It was replaced
by a list of checkboxes
Finally, a first version of read-only possibility has been implemented,
although not fully tested, as it has no direct use yet.